+++ PoliCCTAngular/src/app/ubicacion/ubicacion.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, Inject, ViewChild } from '@angular/core';
 import { Ubicacion } from '../models/ubicaciones.model';
 import { UbicacionService } from '../services/ubicacion.service';
 import { MatTableDataSource } from '@angular/material';
+import { Sede } from '../models/sedes.model';
+import { SedeService } from '../services/sede.service';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 
@@ -13,19 +15,30 @@ import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dial
 export class UbicacionComponent implements OnInit {
 
   @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
+  newUbicacion: Ubicacion ;
   misUbicaciones ; 
-  newUbicacion: Ubicacion;
+  sede: Sede ;
+  misSedes ; 
   dataSource = new MatTableDataSource(this.misUbicaciones);
   No = 0 ;
-  
-  constructor(private UbicacionService: UbicacionService , public dialog: MatDialog ) { 
-    this.getAllUbicaciones();
+    
+  constructor(private UbicacionService: UbicacionService, 
+    private SedeService: SedeService, public dialog: MatDialog ) { 
+     this.getAllUbicaciones();
   }
   ngOnInit() {
     this.dataSource.paginator = this.paginator;
   }  
   getAllUbicaciones(){
     this.UbicacionService.getAllUbicaciones().subscribe(   misUbicacionesObs => {   this.misUbicaciones = misUbicacionesObs;   }   )
+    this.SedeService.getAllSedes().subscribe(   misSedesObs => {   this.misSedes = misSedesObs;   }   )
+    for (let a = 0; a < this.misUbicaciones.length; a++) {
+      for (let b = 0; b < this.misSedes.length; b++) {
+        if ( this.misUbicaciones[a].idSede === this.misSedes[b].idSede ){
+          this.misUbicaciones[a].idSede = this.misSedes[b].nombreSede ; 
+        }
+      }
+    }
     this.dataSource = new MatTableDataSource(this.misUbicaciones);
     this.newUbicacion = new Ubicacion;
     for (let i = 0; i < this.misUbicaciones.length -1 ; i++) {
@@ -41,42 +54,59 @@ export class UbicacionComponent implements OnInit {
   openDialogNuevaUbicacion(): void {
     const dialogRef = this.dialog.open(UbicacionEmergente, {
       width: '300px',
-      data: { Ubicacion }
+      data: { misSedes: this.misSedes , misUbicaciones: this.misUbicaciones }
     });
-    dialogRef.afterClosed().subscribe(result => {
-      this.newUbicacion.nombreUbicacion = result;
-      if (this.newUbicacion.nombreUbicacion === undefined ){
+    dialogRef.afterClosed().subscribe(result => {      
+      if ( result.nombreSede === undefined || result.nombreUbicacion === undefined ){
+        var r = alert('Datos Incompletos');
       }
       else{
         this.newUbicacion.idUbicacion = this.No ; 
+        this.newUbicacion.nombreUbicacion = result.nombreUbicacion ;
+        for (let b = 0; b < this.misSedes.length; b++) {
+          if ( result.nombreSede === this.misSedes[b].nombreSede ){
+            this.newUbicacion.idSede = this.misSedes[b].idSede ;
+          }
+        }
         this.UbicacionService.createNewUbicacion(this.newUbicacion);
         this.getAllUbicaciones();
+        var r = alert('Registro Exitoso');
       }
     });
   }
   openDialogEditarUbicacion(element): void {
     const dialogRef = this.dialog.open(UbicacionEmergente, {
       width: '300px',
-      data: { nombreUbicacion: this.misUbicaciones[element].nombreUbicacion }
+      data: { misSedes: this.misSedes, nombreUbicacion: this.misUbicaciones[element].nombreUbicacion, nombreSede: this.misUbicaciones[element].idSede }
     });
     dialogRef.afterClosed().subscribe(result => {
-      if (result === undefined ){
-      }
-      else{
-        let número = this.misUbicaciones[element].idUbicacion ; 
-        this.eliminarUbicacion(element);
-        this.newUbicacion.idUbicacion = número ;
-        this.newUbicacion.nombreUbicacion = result;
-        this.UbicacionService.createNewUbicacion(this.newUbicacion);
+      if ( result.nombreSede === undefined || result.nombreUbicacion === undefined ||
+        result.nombreSede === null || result.nombreUbicacion === null  ){
+          var r = alert('Datos Incompletos');
+      }else{
+        for (let b = 0; b < this.misSedes.length; b++) {
+          if ( result.nombreSede === this.misSedes[b].nombreSede ){
+            this.newUbicacion.idSede = this.misSedes[b].idSede ;
+          }
+        }
+        this.UbicacionService.editarUbicacion(element, this.newUbicacion.idSede , result.nombreUbicacion);
         this.getAllUbicaciones();
+        var r = alert('Registro Exitoso');
       }
     });
   }
   eliminarUbicacion(element){
-    this.UbicacionService.eliminarUbicacion(element);
-    this.getAllUbicaciones();
+    var r = confirm('¿Esta seguro que desea Eliminar el Registro');
+    if(r === true){
+      this.UbicacionService.eliminarUbicacion(element);
+      this.getAllUbicaciones();
+      var r1 = alert('Registro Eliminado Exitosamente');
+      return true;
+    }else{
+      return false;
+    }
   } 
-  displayedColumns: string[] = ['idUbicacion', 'nombreUbicacion', "star"];
+  displayedColumns: string[] = ['idUbicacion', 'nombreUbicacion', 'Sede', "star"];
   applyFilter(event: Event) {
     const filterValue = (event.target as HTMLInputElement).value;
     this.dataSource.filter = filterValue.trim().toLowerCase();
@@ -88,7 +118,7 @@ export class UbicacionComponent implements OnInit {
 })
 export class UbicacionEmergente {
   constructor(public dialogRef: MatDialogRef<UbicacionEmergente>,
-    @Inject(MAT_DIALOG_DATA)  public data: Ubicacion ) {    }
+    @Inject(MAT_DIALOG_DATA)  public data: { misSedes , misUbicaciones, nombreSede, nombreUbicacion } ){    }
   onNoClick(): void {
     this.dialogRef.close();
   }
